floatをdoubleに

+++ Inc/complex.hpp
@@ -5,14 +5,14 @@
 #include <algorithm>
 
 struct Complex {
-    float real;
-    float im;
-    float abs;
+    double real;
+    double im;
+    double abs;
 
-    Complex() : real(0.0f), im(0.0f), abs(0.0f) {}
-    Complex(float real) : real(real), im(0.0f) { this->abs = fabs(real); }
-    Complex(float real, float im) : real(real), im(im), abs(sqrt(real * real + im * im)) {}
-    Complex(float real, float im, float abs) : real(real), im(im), abs(abs) {}
+    Complex() : real(0.0), im(0.0), abs(0.0) {}
+    Complex(double real) : real(real), im(0.0) { this->abs = fabs(real); }
+    Complex(double real, double im) : real(real), im(im), abs(sqrt(real * real + im * im)) {}
+    Complex(double real, double im, double abs) : real(real), im(im), abs(abs) {}
 
     Complex operator+(Complex b)
     {
@@ -45,17 +45,17 @@ struct Complex {
 
 Complex mid(Complex data[], const int len)
 {
-    float real_list[len], im_list[len];
+    double real_list[len], im_list[len];
     for (int i = 0; i < len; ++i) {
         real_list[i] = data[i].real;
         im_list[i] = data[i].im;
     }
     std::sort(&real_list[0], &real_list[len - 1]);
     std::sort(&im_list[0], &im_list[len - 1]);
-    float real_med = (len % 2 == 0)
+    double real_med = (len % 2 == 0)
                          ? (real_list[len / 2] + real_list[len / 2 - 1]) / 2
                          : real_list[len / 2];
-    float im_med = (len % 2 == 0)
+    double im_med = (len % 2 == 0)
                        ? (im_list[len / 2] + im_list[len / 2 - 1]) / 2
                        : im_list[len / 2];
     return Complex(real_med, im_med);
